requirements.txt
asgiref==3.7.2
boto3==1.34.88
botocore==1.34.88
Django==5.0.2
django-bcrypt==0.9.2
django-cors-headers==4.3.1
django-pandas==0.6.7
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
gunicorn==22.0.0
itsdangerous==2.1.2
jmespath==1.0.1
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pillow==10.2.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
py-bcrypt==0.4
PyJWT==2.8.0
pyotp==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
s3transfer==0.10.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
django-heroku
dj-database-url
